Split network.network into submodules
The file was pretty lonk and complex. Splitting it should make it
more comprehensible.
