+++ src/oemof/network/network/entity.py
@@ -1,13 +1,7 @@
 # -*- coding: utf-8 -*-
 
-"""This package (along with its subpackages) contains the classes used to model
-energy systems. An energy system is modelled as a graph/network of entities
-with very specific constraints on which types of entities are allowed to be
-connected.
-
-This file is part of project oemof (github.com/oemof/oemof). It's copyrighted
-by the contributors recorded in the version control history of the file,
-available from its original location oemof/oemof/network.py
+"""This package contains the abstract entity classes used to model
+energy systems.
 
 SPDX-FileCopyrightText: Stephan Günther <>
 SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
@@ -18,58 +12,10 @@ SPDX-FileCopyrightText: Patrik Schönfeldt <patrik.schoenfeldt@dlr.de>
 SPDX-License-Identifier: MIT
 """
 
-from collections import UserDict
-from collections import namedtuple
-from collections.abc import Mapping
-from collections.abc import MutableMapping
 from functools import total_ordering
 
-
-class Inputs(MutableMapping):
-    """A special helper to map `n1.inputs[n2]` to `n2.outputs[n1]`."""
-
-    def __init__(self, target):
-        self.target = target
-
-    def __getitem__(self, key):
-        return key.outputs.__getitem__(self.target)
-
-    def __delitem__(self, key):
-        return key.outputs.__delitem__(self.target)
-
-    def __setitem__(self, key, value):
-        return key.outputs.__setitem__(self.target, value)
-
-    def __iter__(self):
-        return iter(self.target._in_edges)
-
-    def __len__(self):
-        return self.target._in_edges.__len__()
-
-    def __repr__(self):
-        return repr(
-            "<{0.__module__}.{0.__name__}: {1!r}>".format(
-                type(self), dict(self)
-            )
-        )
-
-
-class Outputs(UserDict):
-    """
-    Helper that intercepts modifications to update `Inputs` symmetrically.
-    """
-
-    def __init__(self, source):
-        self.source = source
-        super().__init__()
-
-    def __delitem__(self, key):
-        key._in_edges.remove(self.source)
-        return super().__delitem__(key)
-
-    def __setitem__(self, key, value):
-        key._in_edges.add(self.source)
-        return super().__setitem__(key, value)
+from .helpers import Inputs
+from .helpers import Outputs
 
 
 @total_ordering
@@ -244,122 +190,6 @@ class Entity:
         return self._outputs
 
 
-EdgeLabel = namedtuple("EdgeLabel", ["input", "output"])
-
-
-class Edge(Entity):
-    """
-    :class:`Bus`es/:class:`Component`s are always connected by an
-    :class:`Edge`.
-
-    :class:`Edge`s connect a single :class:`Node` with another. They are
-    directed and have a (sequence of) value(s) attached to them, so they can
-    be used to represent a flow from a source/an input to a target/an output.
-
-    Parameters
-    ----------
-    input, output: :class:`Bus` or :class:`Component`, optional
-    flow, values: object, optional
-        The (list of) object(s) representing the values flowing from this
-        edge's input into its output. Note that these two names are aliases of
-        each other, so `flow` and `values` are mutually exclusive.
-
-    Note that all of these parameters are also set as attributes with the same
-    name.
-    """
-
-    Label = EdgeLabel
-
-    def __init__(
-        self, input=None, output=None, flow=None, values=None, **kwargs
-    ):
-        if flow is not None and values is not None:
-            raise ValueError(
-                "\n\n`Edge`'s `flow` and `values` keyword arguments are "
-                "aliases of each other,\nso they're mutually exclusive.\n"
-                "You supplied:\n"
-                + "    `flow`  : {}\n".format(flow)
-                + "    `values`: {}\n".format(values)
-                + "Choose one."
-            )
-        super().__init__(label=Edge.Label(input, output))
-        self.values = values if values is not None else flow
-        if input is not None and output is not None:
-            input.outputs[output] = self
-
-    @classmethod
-    def from_object(cls, o):
-        """Creates an `Edge` instance from a single object.
-
-        This method inspects its argument and does something different
-        depending on various cases:
-
-          * If `o` is an instance of `Edge`, `o` is returned unchanged.
-          * If `o` is a `Mapping`, the instance is created by calling
-            `cls(**o)`,
-          * In all other cases, `o` will be used as the `values` keyword
-            argument to `Edge`'s constructor.
-        """
-        if isinstance(o, Edge):
-            return o
-        elif isinstance(o, Mapping):
-            return cls(**o)
-        else:
-            return Edge(values=o)
-
-    @property
-    def flow(self):
-        return self.values
-
-    @flow.setter
-    def flow(self, values):
-        self.values = values
-
-    @property
-    def input(self):
-        return self.label.input
-
-    @input.setter
-    def input(self, i):
-        old_input = self.input
-        self.label = Edge.Label(i, self.label.output)
-        if old_input is None and i is not None and self.output is not None:
-            i.outputs[self.output] = self
-
-    @property
-    def output(self):
-        return self.label.output
-
-    @output.setter
-    def output(self, o):
-        old_output = self.output
-        self.label = Edge.Label(self.label.input, o)
-        if old_output is None and o is not None and self.input is not None:
-            o.inputs[self.input] = self
-
-
-class Node(Entity):
-    pass
-
-
-class Bus(Node):
-    pass
-
-
-class Component(Node):
-    pass
-
-
-class Sink(Component):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-
-class Source(Component):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-
-class Transformer(Component):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
+# Need to import in the end to avoid circular import.
+# (Above, a string representation is used to allow this late import.)
+from .edge import Edge  # noqa: F401, E402
