+++ src/oemof/network/network/nodes.py
@@ -1,0 +1,41 @@
+# -*- coding: utf-8 -*-
+"""This package contains the differnt types of Node for
+modelling an energy system graph.
+
+SPDX-FileCopyrightText: Stephan Günther <>
+SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
+SPDX-FileCopyrightText: Simon Hilpert <>
+SPDX-FileCopyrightText: Cord Kaldemeyer <>
+SPDX-FileCopyrightText: Patrik Schönfeldt <patrik.schoenfeldt@dlr.de>
+
+SPDX-License-Identifier: MIT
+"""
+
+from .entity import Entity
+
+
+class Node(Entity):
+    pass
+
+
+class Bus(Node):
+    pass
+
+
+class Component(Node):
+    pass
+
+
+class Sink(Component):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+
+class Source(Component):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+
+class Transformer(Component):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
